+++ src/main/java/gregtech/common/tileentities/machines/multi/MTEIndustrialElectrolyzer.java
@@ -50,8 +50,8 @@ public class MTEIndustrialElectrolyzer extends MTEExtendedPowerMultiBlockBase<MT
     private static final int OFFSET_Z = 1;
 
     private static final int PARALLEL_PER_TIER = 4;
-    private static final float SPEED = 2.8f;
-    private static final float EU_EFFICIENCY = 0.9f;
+    private static final double SPEED = 2.8D;
+    private static final double EU_EFFICIENCY = 0.9D;
 
     public MTEIndustrialElectrolyzer(final int aID, final String aName, final String aNameRegional) {
         super(aID, aName, aNameRegional);
@@ -146,7 +146,7 @@ public class MTEIndustrialElectrolyzer extends MTEExtendedPowerMultiBlockBase<MT
     protected MultiblockTooltipBuilder createTooltip() {
         MultiblockTooltipBuilder tt = new MultiblockTooltipBuilder();
         tt.addMachineType("Electrolyzer")
-            .addBulkMachineInfo(PARALLEL_PER_TIER, SPEED, EU_EFFICIENCY)
+            .addBulkMachineInfo(PARALLEL_PER_TIER, (float) SPEED, (float) EU_EFFICIENCY)
             .addPollutionAmount(getPollutionPerSecond(null))
             .beginStructureBlock(5, 5, 5, false)
             .addController("Front center, 3rd layer")
@@ -166,7 +166,7 @@ public class MTEIndustrialElectrolyzer extends MTEExtendedPowerMultiBlockBase<MT
 
     @Override
     protected ProcessingLogic createProcessingLogic() {
-        return new ProcessingLogic().setSpeedBonus(1F / SPEED)
+        return new ProcessingLogic().setSpeedBonus(1.0D / SPEED)
             .setEuModifier(EU_EFFICIENCY)
             .setMaxParallelSupplier(this::getTrueParallel);
     }
